go clear the code

+++ includes/Server.hpp
@@ -42,7 +42,7 @@ private:
 	std::string 				server_name;
 	std::string					root;
 	std::vector<std::string>	errors;
-	int 						client_body_size;
+	size_t 						client_body_size;
 	std::vector<Location> 		locations;
 
     sockaddr_in         		addr;
@@ -74,7 +74,7 @@ public:
 	std::string						getServerName(void) const;
 	std::string						getRoot(void) const;
 	std::vector<std::string>		&getErrors(void);
-	int								getClientBodySize(void) const;
+	size_t							getClientBodySize(void) const;
 	std::vector<Location>			*getLocations(void);
 
 	// SETTERS
